move code to legacy directory
