+++ src/index.ts
@@ -1,1 +1,2 @@
-export * from './public-api'
+export * from "./lifecycle"
+export * from "./utils"
