+++ src/legacy/decorators/inject.decorator.ts
@@ -1,5 +1,5 @@
 import { addDependency, InjectionToken, Token } from "../container"
-import { Constructor } from "../utils"
+import { Constructor } from "../../utils"
 
 /**
  * @author kuro <kuro@kuroi.io>
