+++ src/legacy/index.ts
@@ -1,0 +1,2 @@
+export * from "./container";
+export * from "./decorators";
